requirements.txt
Flask
requests
gunicorn
pytz
python-docx